refactor(core): split feats and proficiencies into leaf modules
Keep public API via thin facades; split menu character tests by flow.
